converted note sidebar to the new tabbed structure to reduce scrolling
