+++ DLPrototype/Views/Notes/Sidebars/NoteDashboardSidebar.swift
@@ -9,14 +9,56 @@
 import SwiftUI
 
 struct NoteDashboardSidebar: View {
+    @State private var tabs: [ToolbarButton] = []
+    @State private var searching: Bool = false
+
     var body: some View {
         ScrollView(showsIndicators: false) {
             VStack(alignment: .leading, spacing: 5) {
-                NoteSearchWidget()
-                FavouriteNotesWidget()
+                FancyGenericToolbar(
+                    buttons: tabs,
+                    standalone: true,
+                    location: .sidebar,
+                    mode: .compact
+                )
             }
             Spacer()
         }
         .padding()
+        .onAppear(perform: createToolbar)
+    }
+}
+
+extension NoteDashboardSidebar {
+    private func createToolbar() -> Void {
+        tabs = [
+            ToolbarButton(
+                id: 0,
+                helpText: "Notes",
+                icon: "note.text",
+                labelText: "Notes",
+                contents: AnyView(NotesWidget())
+            ),
+            ToolbarButton(
+                id: 1,
+                helpText: "Favourites notes",
+                icon: "star.fill",
+                labelText: "Favourite Notes",
+                contents: AnyView(NotesWidget(favouritesOnly: true))
+            ),
+            ToolbarButton(
+                id: 3,
+                helpText: "Search",
+                icon: "magnifyingglass",
+                labelText: "Search",
+                contents: AnyView(
+                    VStack(alignment: .leading) {
+                        FindDashboard(searching: $searching, location: .sidebar)
+                    }
+                    .padding(8)
+                    .background(Theme.base.opacity(0.2))
+                )
+            )
+        ]
     }
 }
